feat: separate dask and non-dask function, create new repertory for distributed computing
